+++ engines/src/mech/engine_super_elastic_cpu.hpp
@@ -100,7 +100,7 @@ public:
   std::vector<index_t> stoich_coef;
 
   // number of variables per jacobian matrix block
-  const static uint8_t N_VARS_SQ = N_VARS * N_VARS;
+  const static uint16_t N_VARS_SQ = N_VARS * N_VARS;
 
   uint8_t get_n_vars() const { return N_VARS; };
   uint8_t get_n_ops() const  { return N_OPS; };
